refactor: PostService와 ThumbsService의 양방향 연관을 중간 interface를 두어 thumbs -> post만 되도록 수정함
