+++ backend/src/main/java/edonymyeon/backend/post/application/dto/ThumbsStatusInPostResponse.java
@@ -1,4 +1,4 @@
-package edonymyeon.backend.thumbs.dto;
+package edonymyeon.backend.post.application.dto;
 
 public record ThumbsStatusInPostResponse(boolean isUp, boolean isDown) {
